+++ moirai-async/src/net/tests/mod.rs
@@ -1,5 +1,7 @@
 #![cfg_attr(test, allow(clippy::unwrap_used, reason = "test scope"))]
 
+mod connect;
+
 use super::*;
 use crate::executor::AsyncExecutor;
 use crate::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};